+++ actions/src/robocorp/actions/__init__.py
@@ -8,7 +8,7 @@ from ._protocols import IAction, Status
 from ._request import Request
 from ._secret import Secret
 
-__version__ = "0.1.3"
+__version__ = "0.2.0"
 version_info = [int(x) for x in __version__.split(".")]